Support 32-bit machines
The crate refused to compile unless `target_pointer_width` was 64, via a
`compile_error!` in `lib.rs` and `#![cfg(target_pointer_width = "64")]` on
most modules. Removing them exposed only a handful of real dependencies on
the pointer width, all of which are fixed here.

- `broadword`: the 64-bit step constants were typed `usize`, so they
  overflowed at compile time on a 32-bit target. They are now `u64`, along
  with the helpers that consume them. On a 64-bit machine the types are
  identical, so nothing changes there.
- `Rank9SelIndex::block_rank_pairs`: it packs seven 9-bit subranks into one
  word, which does not fit in a 32-bit `usize`. It is now `Vec<u64>`.
- `DArrayIndex::block_inventory`: bit positions stored as `isize` would wrap
  into the negative overflow marker beyond 2^31 bits on a 32-bit machine.
  It is now `Vec<i64>`.
- `EliasFanoBuilder::push`: `val as usize >> low_len` cast before shifting
  and thus truncated the value. It now shifts before casting.
- `EliasFano::delta`: it computed the delta in `usize`, which overflows when
  `low_len >= 32`. It now computes in `u64`.

`usize` and `isize` are serialized as fixed 64-bit little-endian integers
instead of native-width ones, so the format no longer depends on the machine.
The bytes are unchanged from those written by 0.9.1 on a 64-bit machine, so
existing data stay readable; deserializing a value that does not fit in
`usize` of the machine now fails with `SucdsError::InvalidArgument`.

Verified that the serialized bytes and the query results are identical
between 0.9.1 on 64-bit, this commit on 64-bit, and this commit on a 32-bit
target, and that all unit tests pass on wasm32-wasip1.

CI gains a `test-32bit` job that builds and runs the tests for
i686-unknown-linux-gnu (i686 binaries run natively on the x86_64 runner, so
the tests really execute with a 32-bit `usize`), and the no_std job also
builds for thumbv7em-none-eabi to cover no_std and a 32-bit pointer width at
once.

+++ src/broadword.rs
@@ -1,44 +1,43 @@
 //! Broadword tools.
-#![cfg(target_pointer_width = "64")]
 #![allow(dead_code)]
 
 #[cfg(feature = "intrinsics")]
 use crate::intrinsics;
 
 pub(crate) const ONES_STEP_4: u64 = 0x1111111111111111;
-pub(crate) const ONES_STEP_8: usize = 0x0101010101010101;
-pub(crate) const ONES_STEP_9: usize =
+pub(crate) const ONES_STEP_8: u64 = 0x0101010101010101;
+pub(crate) const ONES_STEP_9: u64 =
     (1 << 0) | (1 << 9) | (1 << 18) | (1 << 27) | (1 << 36) | (1 << 45) | (1 << 54);
-pub(crate) const MSBS_STEP_8: usize = 0x80 * ONES_STEP_8;
-pub(crate) const MSBS_STEP_9: usize = 0x100 * ONES_STEP_9;
-pub(crate) const INV_COUNT_STEP_9: usize =
+pub(crate) const MSBS_STEP_8: u64 = 0x80 * ONES_STEP_8;
+pub(crate) const MSBS_STEP_9: u64 = 0x100 * ONES_STEP_9;
+pub(crate) const INV_COUNT_STEP_9: u64 =
     (1 << 54) | (2 << 45) | (3 << 36) | (4 << 27) | (5 << 18) | (6 << 9) | 7;
 
 #[inline(always)]
-pub(crate) const fn leq_step_8(x: usize, y: usize) -> usize {
+pub(crate) const fn leq_step_8(x: u64, y: u64) -> u64 {
     ((((y | MSBS_STEP_8) - (x & !MSBS_STEP_8)) ^ (x ^ y)) & MSBS_STEP_8) >> 7
 }
 
 #[inline(always)]
-pub(crate) const fn uleq_step_8(x: usize, y: usize) -> usize {
+pub(crate) const fn uleq_step_8(x: u64, y: u64) -> u64 {
     (((((y | MSBS_STEP_8) - (x & !MSBS_STEP_8)) ^ (x ^ y)) ^ (x & !y)) & MSBS_STEP_8) >> 7
 }
 
 #[inline(always)]
-pub(crate) const fn uleq_step_9(x: usize, y: usize) -> usize {
+pub(crate) const fn uleq_step_9(x: u64, y: u64) -> u64 {
     (((((y | MSBS_STEP_9) - (x & !MSBS_STEP_9)) | (x ^ y)) ^ (x & !y)) & MSBS_STEP_9) >> 8
 }
 
 #[inline(always)]
-pub(crate) const fn byte_counts(mut x: u64) -> usize {
+pub(crate) const fn byte_counts(mut x: u64) -> u64 {
     x = x - ((x & (0xa * ONES_STEP_4)) >> 1);
     x = (x & (3 * ONES_STEP_4)) + ((x >> 2) & (3 * ONES_STEP_4));
-    ((x + (x >> 4)) & (0x0f * ONES_STEP_8 as u64)) as usize
+    (x + (x >> 4)) & (0x0f * ONES_STEP_8)
 }
 
 #[inline(always)]
 pub(crate) const fn bytes_sum(x: u64) -> usize {
-    ((ONES_STEP_8 as u64).wrapping_mul(x) >> 56) as usize
+    (ONES_STEP_8.wrapping_mul(x) >> 56) as usize
 }
 
 /// Counts the number of set bits.
@@ -56,7 +55,7 @@ pub(crate) const fn bytes_sum(x: u64) -> usize {
 pub const fn popcount(x: u64) -> usize {
     #[cfg(not(feature = "intrinsics"))]
     {
-        bytes_sum(byte_counts(x) as u64)
+        bytes_sum(byte_counts(x))
     }
     #[cfg(feature = "intrinsics")]
     {
@@ -83,19 +82,19 @@ pub const fn select_in_word(x: u64, k: usize) -> Option<usize> {
         return None;
     }
     let byte_sums = ONES_STEP_8.wrapping_mul(byte_counts(x));
-    let k_step_8 = k * ONES_STEP_8;
+    let k_step_8 = k as u64 * ONES_STEP_8;
     let geq_k_step_8 = ((k_step_8 | MSBS_STEP_8) - byte_sums) & MSBS_STEP_8;
     let place = {
         #[cfg(feature = "intrinsics")]
         {
-            popcount(geq_k_step_8 as u64) * 8
+            popcount(geq_k_step_8) * 8
         }
         #[cfg(not(feature = "intrinsics"))]
         {
-            ((geq_k_step_8 >> 7).wrapping_mul(ONES_STEP_8) >> 53) & !0x7
+            (((geq_k_step_8 >> 7).wrapping_mul(ONES_STEP_8) >> 53) & !0x7) as usize
         }
     };
-    let byte_rank = k - (((byte_sums << 8) >> place) & 0xFF);
+    let byte_rank = k - (((byte_sums << 8) >> place) & 0xFF) as usize;
     let sel = place + SELECT_IN_BYTE[((x >> place) as usize & 0xFF) | (byte_rank << 8)] as usize;
     Some(sel)
 }
